fix: resolve OAuth on merged IR instead of per-converter
