+++ packages/cli/api-importers/openapi-to-ir/src/3.1/OpenAPIConverter.ts
@@ -1,12 +1,4 @@
-import { RawSchemas, visitRawAuthSchemeDeclaration } from "@fern-api/fern-definition-schema";
-import {
-    AuthScheme,
-    FernIr,
-    HttpMethod,
-    IntermediateRepresentation,
-    OAuthConfiguration,
-    RequestPropertyValue
-} from "@fern-api/ir-sdk";
+import { IntermediateRepresentation } from "@fern-api/ir-sdk";
 import { constructHttpPath, convertApiAuth, convertEnvironments } from "@fern-api/ir-utils";
 import {
     AbstractSpecConverter,
@@ -61,8 +53,6 @@ export class OpenAPIConverter extends AbstractSpecConverter<OpenAPIConverterCont
         this.convertWebhooks();
 
         const { endpointLevelServers, errors } = this.convertPaths();
-
-        this.resolveOAuthEndpointReferences();
 
         this.addErrorsToIr(errors);
 
@@ -349,349 +339,6 @@ export class OpenAPIConverter extends AbstractSpecConverter<OpenAPIConverterCont
 
         this.context.spec.security = [securityRequirement];
         this.removeEndpointSpecificAuth();
-    }
-
-    private resolveOAuthEndpointReferences(): void {
-        if (!this.context.authOverrides?.["auth-schemes"] || !this.ir.auth) {
-            return;
-        }
-
-        const authSchemes = this.context.authOverrides["auth-schemes"];
-        const updatedSchemes: AuthScheme[] = [];
-
-        for (const scheme of this.ir.auth.schemes) {
-            // Only replace bearer schemes that were placeholders for OAuth
-            if (scheme.type !== "bearer") {
-                updatedSchemes.push(scheme);
-                continue;
-            }
-
-            const rawDeclaration = authSchemes[scheme.key];
-            if (rawDeclaration == null) {
-                updatedSchemes.push(scheme);
-                continue;
-            }
-
-            const resolvedScheme = visitRawAuthSchemeDeclaration<AuthScheme | undefined>(rawDeclaration, {
-                header: () => undefined,
-                basic: () => undefined,
-                tokenBearer: () => undefined,
-                inferredBearer: () => undefined,
-                oauth: (oauthScheme) => this.resolveOAuthScheme({ key: scheme.key, oauthScheme })
-            });
-
-            updatedSchemes.push(resolvedScheme ?? scheme);
-        }
-
-        this.ir.auth = {
-            ...this.ir.auth,
-            schemes: updatedSchemes
-        };
-    }
-
-    private resolveOAuthScheme({
-        key,
-        oauthScheme
-    }: {
-        key: string;
-        oauthScheme: RawSchemas.OAuthSchemeSchema;
-    }): AuthScheme | undefined {
-        if (oauthScheme.type !== "client-credentials") {
-            return undefined;
-        }
-
-        const tokenEndpointConfig = oauthScheme["get-token"];
-        const parsed = this.parseEndpointReference(tokenEndpointConfig.endpoint);
-        if (parsed == null) {
-            return undefined;
-        }
-
-        const { method, path } = parsed;
-        const resolved = this.findEndpointInIr({ method, path });
-        if (resolved == null) {
-            this.context.errorCollector.collect({
-                message: `Could not resolve OAuth token endpoint: ${tokenEndpointConfig.endpoint}`,
-                path: ["auth-schemes"]
-            });
-            return undefined;
-        }
-
-        const { endpoint, serviceId, subpackageId } = resolved;
-
-        const requestProperties = tokenEndpointConfig["request-properties"];
-        const responseProperties = tokenEndpointConfig["response-properties"];
-
-        const clientIdProp = this.findRequestBodyProperty({
-            endpoint,
-            wireValue: this.extractPropertyName(requestProperties?.["client-id"]) ?? "client_id"
-        });
-        const clientSecretProp = this.findRequestBodyProperty({
-            endpoint,
-            wireValue: this.extractPropertyName(requestProperties?.["client-secret"]) ?? "client_secret"
-        });
-        const accessTokenProp = this.findResponseProperty({
-            endpoint,
-            wireValue: this.extractPropertyName(responseProperties?.["access-token"]) ?? "access_token"
-        });
-
-        if (clientIdProp == null || clientSecretProp == null || accessTokenProp == null) {
-            this.context.errorCollector.collect({
-                message: `Could not resolve OAuth token endpoint properties for: ${tokenEndpointConfig.endpoint}`,
-                path: ["auth-schemes"]
-            });
-            return undefined;
-        }
-
-        const expiresInProp = this.findResponseProperty({
-            endpoint,
-            wireValue: this.extractPropertyName(responseProperties?.["expires-in"]) ?? "expires_in"
-        });
-        const refreshTokenProp = this.findResponseProperty({
-            endpoint,
-            wireValue: this.extractPropertyName(responseProperties?.["refresh-token"]) ?? "refresh_token"
-        });
-
-        let refreshEndpoint: FernIr.OAuthRefreshEndpoint | undefined;
-        const refreshTokenConfig = oauthScheme["refresh-token"];
-        if (refreshTokenConfig != null) {
-            refreshEndpoint = this.resolveRefreshEndpoint({ refreshTokenConfig });
-        }
-
-        return AuthScheme.oauth({
-            key,
-            docs: oauthScheme.docs,
-            configuration: OAuthConfiguration.clientCredentials({
-                clientIdEnvVar: oauthScheme["client-id-env"],
-                clientSecretEnvVar: oauthScheme["client-secret-env"],
-                tokenPrefix: oauthScheme["token-prefix"],
-                tokenHeader: oauthScheme["token-header"],
-                scopes: oauthScheme.scopes,
-                tokenEndpoint: {
-                    endpointReference: {
-                        endpointId: endpoint.id,
-                        serviceId,
-                        subpackageId
-                    },
-                    requestProperties: {
-                        clientId: clientIdProp,
-                        clientSecret: clientSecretProp,
-                        scopes: undefined,
-                        customProperties: undefined
-                    },
-                    responseProperties: {
-                        accessToken: accessTokenProp,
-                        expiresIn: expiresInProp,
-                        refreshToken: refreshTokenProp
-                    }
-                },
-                refreshEndpoint
-            })
-        });
-    }
-
-    private resolveRefreshEndpoint({
-        refreshTokenConfig
-    }: {
-        refreshTokenConfig: RawSchemas.OAuthRefreshTokenEndpointSchema;
-    }): FernIr.OAuthRefreshEndpoint | undefined {
-        const parsed = this.parseEndpointReference(refreshTokenConfig.endpoint);
-        if (parsed == null) {
-            return undefined;
-        }
-
-        const resolved = this.findEndpointInIr(parsed);
-        if (resolved == null) {
-            return undefined;
-        }
-
-        const { endpoint, serviceId, subpackageId } = resolved;
-        const responseProperties = refreshTokenConfig["response-properties"];
-        const requestProperties = refreshTokenConfig["request-properties"];
-
-        const refreshTokenReqProp = this.findRequestBodyProperty({
-            endpoint,
-            wireValue: this.extractPropertyName(requestProperties?.["refresh-token"]) ?? "refresh_token"
-        });
-        const accessTokenProp = this.findResponseProperty({
-            endpoint,
-            wireValue: this.extractPropertyName(responseProperties?.["access-token"]) ?? "access_token"
-        });
-
-        if (refreshTokenReqProp == null || accessTokenProp == null) {
-            return undefined;
-        }
-
-        const expiresInProp = this.findResponseProperty({
-            endpoint,
-            wireValue: this.extractPropertyName(responseProperties?.["expires-in"]) ?? "expires_in"
-        });
-        const refreshTokenRespProp = this.findResponseProperty({
-            endpoint,
-            wireValue: this.extractPropertyName(responseProperties?.["refresh-token"]) ?? "refresh_token"
-        });
-
-        return {
-            endpointReference: {
-                endpointId: endpoint.id,
-                serviceId,
-                subpackageId
-            },
-            requestProperties: {
-                refreshToken: refreshTokenReqProp
-            },
-            responseProperties: {
-                accessToken: accessTokenProp,
-                expiresIn: expiresInProp,
-                refreshToken: refreshTokenRespProp
-            }
-        };
-    }
-
-    private parseEndpointReference(endpointRef: string): { method: HttpMethod; path: string } | undefined {
-        const parts = endpointRef.trim().split(/\s+/);
-        if (parts.length !== 2) {
-            return undefined;
-        }
-        const [methodStr, path] = parts;
-        const method = this.parseHttpMethod(methodStr ?? "");
-        if (method == null || path == null) {
-            return undefined;
-        }
-        return { method, path };
-    }
-
-    private parseHttpMethod(method: string): HttpMethod | undefined {
-        switch (method.toUpperCase()) {
-            case "GET":
-                return HttpMethod.Get;
-            case "POST":
-                return HttpMethod.Post;
-            case "PUT":
-                return HttpMethod.Put;
-            case "PATCH":
-                return HttpMethod.Patch;
-            case "DELETE":
-                return HttpMethod.Delete;
-            default:
-                return undefined;
-        }
-    }
-
-    private findEndpointInIr({ method, path }: { method: HttpMethod; path: string }):
-        | {
-              endpoint: FernIr.HttpEndpoint;
-              serviceId: string;
-              subpackageId: string | undefined;
-          }
-        | undefined {
-        const normalizedPath = path.replace(/\/+$/, "");
-
-        for (const [serviceId, service] of Object.entries(this.ir.services)) {
-            for (const endpoint of service.endpoints) {
-                if (endpoint.method !== method) {
-                    continue;
-                }
-
-                const endpointPath = this.reconstructPath(endpoint.fullPath);
-                if (endpointPath === normalizedPath) {
-                    const subpackageId = this.findSubpackageForService(serviceId);
-                    return { endpoint, serviceId, subpackageId };
-                }
-            }
-        }
-        return undefined;
-    }
-
-    private reconstructPath(httpPath: FernIr.HttpPath): string {
-        let result = httpPath.head;
-        for (const part of httpPath.parts) {
-            result += `{${part.pathParameter}}${part.tail}`;
-        }
-        return result.replace(/\/+$/, "");
-    }
-
-    private findSubpackageForService(serviceId: string): string | undefined {
-        for (const [subpackageId, subpackage] of Object.entries(this.ir.subpackages)) {
-            if (subpackage.service === serviceId) {
-                return subpackageId;
-            }
-        }
-        return undefined;
-    }
-
-    private extractPropertyName(value: string | undefined): string | undefined {
-        if (value == null) {
-            return undefined;
-        }
-        // Handle $request.propertyName or $response.propertyName format
-        const match = value.match(/^\$(?:request|response)\.(.+)$/);
-        return match?.[1] ?? value;
-    }
-
-    private findRequestBodyProperty({
-        endpoint,
-        wireValue
-    }: {
-        endpoint: FernIr.HttpEndpoint;
-        wireValue: string;
-    }): FernIr.RequestProperty | undefined {
-        if (endpoint.requestBody == null) {
-            return undefined;
-        }
-
-        if (endpoint.requestBody.type === "inlinedRequestBody") {
-            for (const prop of endpoint.requestBody.properties) {
-                if (prop.name.wireValue === wireValue) {
-                    return {
-                        propertyPath: undefined,
-                        property: RequestPropertyValue.body({
-                            name: prop.name,
-                            valueType: prop.valueType,
-                            availability: prop.availability,
-                            docs: prop.docs,
-                            propertyAccess: prop.propertyAccess,
-                            v2Examples: prop.v2Examples
-                        })
-                    };
-                }
-            }
-        }
-
-        return undefined;
-    }
-
-    private findResponseProperty({
-        endpoint,
-        wireValue
-    }: {
-        endpoint: FernIr.HttpEndpoint;
-        wireValue: string;
-    }): FernIr.ResponseProperty | undefined {
-        if (endpoint.response?.body == null) {
-            return undefined;
-        }
-
-        if (endpoint.response.body.type === "json") {
-            const jsonResponse = endpoint.response.body;
-            if (jsonResponse.value.responseBodyType.type === "named") {
-                const typeId = jsonResponse.value.responseBodyType.typeId;
-                if (typeId != null) {
-                    const typeDecl = this.ir.types[typeId];
-                    if (typeDecl?.shape.type === "object") {
-                        for (const prop of typeDecl.shape.properties) {
-                            if (prop.name.wireValue === wireValue) {
-                                return {
-                                    propertyPath: undefined,
-                                    property: prop
-                                };
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        return undefined;
     }
 
     private removeEndpointSpecificAuth(): void {
