+++ packages/cli/workspace/lazy-fern-workspace/src/OSSWorkspace.ts
@@ -20,7 +20,11 @@ import { IntermediateRepresentation, serialization } from "@fern-api/ir-sdk";
 import { mergeIntermediateRepresentation } from "@fern-api/ir-utils";
 import { OpenApiIntermediateRepresentation } from "@fern-api/openapi-ir";
 import { type ParseOpenAPIOptions, parse } from "@fern-api/openapi-ir-parser";
-import { OpenAPI3_1Converter, OpenAPIConverterContext3_1 } from "@fern-api/openapi-to-ir";
+import {
+    OpenAPI3_1Converter,
+    OpenAPIConverterContext3_1,
+    resolveOAuthEndpointReferences
+} from "@fern-api/openapi-to-ir";
 import { OpenRPCConverter, OpenRPCConverterContext3_1 } from "@fern-api/openrpc-to-ir";
 import { TaskContext } from "@fern-api/task-context";
 import { ErrorCollector } from "@fern-api/v3-importer-commons";
@@ -479,6 +483,13 @@ export class OSSWorkspace extends BaseOpenAPIWorkspace {
         if (mergedIr === undefined) {
             throw new Error("Failed to generate intermediate representation");
         }
+
+        // Resolve OAuth endpoint references after all specs have been merged,
+        // because the token endpoint may be in a different spec than the auth scheme.
+        if (authOverrides != null) {
+            resolveOAuthEndpointReferences({ ir: mergedIr, authOverrides });
+        }
+
         return mergedIr;
     }
